requirements.txt

aniso8601==7.0.0
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==3.2.9
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-js-asset==1.2.2
django-on-heroku==1.1.2
djangorestframework==3.12.4
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
promise==2.3
psycopg2-binary==2.9.2
pycodestyle==2.8.0
python-dotenv==0.19.2
pytz==2021.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
whitenoise==5.3.0
